Fixed layer order glitch with tree hit particle

+++ src/sprites.py
@@ -1,5 +1,6 @@
 from .settings import *
 from .timer import Timer
+from .support import generate_particle_surf
 from random import randint, choice
 
 
@@ -66,13 +67,27 @@ class Tree(CollideableSprite):
     def __init__(self, pos, surf, groups, name, apple_surf, stump_surf):
         super().__init__(pos, surf, groups, (30 * SCALE_FACTOR, 20 * SCALE_FACTOR))
         self.name = name
+        self.part_surf = generate_particle_surf(self.image)
         self.apple_surf = apple_surf
         self.stump_surf = stump_surf
         self.health = 5
+        self.timer = Timer(300, func=self.unhit)
         self.hitbox = None
+        self.was_hit = False
         self.alive = True
         self.apple_sprites = pygame.sprite.Group()
         self.create_fruit()
+
+    def unhit(self):
+        self.was_hit = False
+        if self.health < 0:
+            self.image = self.stump_surf
+            if self.alive:
+                self.rect = self.image.get_frect(midbottom=self.rect.midbottom)
+                self.hitbox = self.rect.inflate(-10, -self.rect.height * 0.6)
+                self.alive = False
+        elif self.health >= 0 and self.alive:
+            self.image = self.surf
 
     def create_fruit(self):
         for pos in APPLE_POS['default']:
@@ -81,20 +96,23 @@ class Tree(CollideableSprite):
                 y = pos[1] + self.rect.top
                 Sprite((x, y), self.apple_surf, (self.apple_sprites, self.groups()[0]), LAYERS['fruit'])
 
+    def update(self, dt):
+        self.timer.update()
+
     def hit(self, entity):
+        if self.was_hit:
+            return
+        self.was_hit = True
         self.health -= 1
         # remove an apple
         if len(self.apple_sprites.sprites()) > 0:
             random_apple = choice(self.apple_sprites.sprites())
             random_apple.kill()
             entity.add_resource('apple')
-        if self.health <= 0 and self.alive:
-            print('x')
-            self.image = self.stump_surf
-            self.rect = self.image.get_frect(midbottom=self.rect.midbottom)
-            self.hitbox = self.rect.inflate(-10, -self.rect.height * 0.6)
-            self.alive = False
-            entity.add_resource('wood', 5)
+        if self.health < 0 and self.alive:
+            entity.add_resource("wood", 5)
+        self.image = generate_particle_surf(self.image)
+        self.timer.activate()
 
 
 class AnimatedSprite(Sprite):
